deleted unused code, renamed new folder
