+++ iris-mpc-cpu/src/lib.rs
@@ -1,10 +1,7 @@
 pub mod database_generators;
 pub(crate) mod execution;
 pub mod hawkers;
-pub(crate) mod networks;
-pub(crate) mod next_gen_network;
-pub mod next_gen_protocol;
+pub(crate) mod network;
 pub mod prelude;
-pub(crate) mod protocol;
+pub mod protocol;
 pub(crate) mod shares;
-pub(crate) mod utils;
