--- iris-mpc-cpu/src/networks/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod network_trait;
-pub(crate) mod test_network;
