--- iris-mpc-cpu/src/next_gen_protocol/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod binary;
-pub mod ng_worker;
