+++ iris-mpc-cpu/src/prelude.rs
@@ -1,7 +1,1 @@
-pub use super::{
-    networks::{
-        network_trait::{NetworkEstablisher, NetworkTrait},
-        test_network::{PartyTestNetwork, TestNetwork3p, TestNetworkEstablisher},
-    },
-    protocol::iris_worker::IrisWorker,
-};
+
