+++ iris-mpc-cpu/src/protocol/binary.rs
@@ -1,7 +1,6 @@
-use super::iris_worker::IrisWorker;
 use crate::{
-    networks::network_trait::NetworkTrait,
-    protocol::iris_worker::{A, A_BITS, B_BITS},
+    execution::session::{Session, SessionHandles},
+    network::value::NetworkValue,
     shares::{
         bit::Bit,
         int_ring::IntRing2k,
@@ -9,663 +8,615 @@ use crate::{
         share::Share,
         vecshare::{SliceShare, VecShare},
     },
-    utils::Utils,
 };
 use eyre::{eyre, Error};
-use iris_mpc_common::id::PartyID;
 use num_traits::{One, Zero};
 use rand::{distributions::Standard, prelude::Distribution, Rng};
 use std::ops::SubAssign;
 
-impl<N: NetworkTrait> IrisWorker<N> {
-    pub(crate) fn mul_lift_2k<T: IntRing2k, const K: u64>(val: &Share<T>) -> Share<u32>
-    where
-        u32: From<T>,
-    {
-        let a = (u32::from(val.a.0)) << K;
-        let b = (u32::from(val.b.0)) << K;
-        Share::new(RingElement(a), RingElement(b))
+pub(crate) fn transposed_padded_len(len: usize) -> usize {
+    let padded_len = (len + 63) / 64;
+    padded_len * 64
+}
+
+pub(crate) fn a2b_pre<T: IntRing2k>(
+    session: &Session,
+    x: Share<T>,
+) -> eyre::Result<(Share<T>, Share<T>, Share<T>)> {
+    let (a, b) = x.get_ab();
+
+    let mut x1 = Share::zero();
+    let mut x2 = Share::zero();
+    let mut x3 = Share::zero();
+
+    match session.own_role()?.zero_based() {
+        0 => {
+            x1.a = a;
+            x3.b = b;
+        }
+        1 => {
+            x2.a = a;
+            x1.b = b;
+        }
+        2 => {
+            x3.a = a;
+            x2.b = b;
+        }
+        _ => {
+            return Err(eyre!(
+                "Cannot deal with roles that have index outside of the set [0, 1, 2]"
+            ))
+        }
+    }
+    Ok((x1, x2, x3))
+}
+
+pub(crate) fn transposed_pack_xor_assign<T: IntRing2k>(x1: &mut [VecShare<T>], x2: &[VecShare<T>]) {
+    let len = x1.len();
+    debug_assert_eq!(len, x2.len());
+
+    for (x1, x2) in x1.iter_mut().zip(x2.iter()) {
+        *x1 ^= x2.as_slice();
+    }
+}
+
+pub(crate) fn transposed_pack_xor<T: IntRing2k>(
+    x1: &[VecShare<T>],
+    x2: &[VecShare<T>],
+) -> Vec<VecShare<T>> {
+    let len = x1.len();
+    debug_assert_eq!(len, x2.len());
+
+    let mut res = Vec::with_capacity(len);
+    for (x1, x2) in x1.iter().zip(x2.iter()) {
+        res.push(x1.as_slice() ^ x2.as_slice());
+    }
+    res
+}
+
+pub(crate) async fn and_many_send(
+    session: &mut Session,
+    a: SliceShare<'_, u64>,
+    b: SliceShare<'_, u64>,
+) -> Result<Vec<RingElement<u64>>, Error>
+where
+    Standard: Distribution<u64>,
+{
+    if a.len() != b.len() {
+        return Err(eyre!("InvalidSize in and_many_send"));
+    }
+    let mut shares_a = Vec::with_capacity(a.len());
+    for (a_, b_) in a.iter().zip(b.iter()) {
+        let rand = session.prf_as_mut().gen_binary_zero_share::<u64>();
+        let mut c = a_ & b_;
+        c ^= rand;
+        shares_a.push(c);
     }
 
-    pub(crate) fn mul_lift_2k_many<T: IntRing2k, const K: u64>(vals: SliceShare<T>) -> VecShare<u32>
-    where
-        u32: From<T>,
-    {
-        VecShare::new_vec(
-            vals.iter()
-                .map(|val| Self::mul_lift_2k::<T, K>(val))
-                .collect(),
+    let next_party = session.next_identity()?;
+    let network = session.network().clone();
+    let sid = session.session_id();
+    let message = shares_a.clone();
+    network
+        .send(
+            NetworkValue::VecRing64(message).to_network(),
+            &next_party,
+            &sid,
         )
-    }
+        .await?;
+    Ok(shares_a)
+}
 
-    pub(crate) fn a2b_pre<T: IntRing2k>(&self, x: Share<T>) -> (Share<T>, Share<T>, Share<T>) {
-        let (a, b) = x.get_ab();
+pub(crate) async fn and_many_receive(
+    session: &mut Session,
+) -> Result<Vec<RingElement<u64>>, Error> {
+    let network = session.network().clone();
+    let sid = session.session_id();
+    let prev_party = session.prev_identity()?;
 
-        let mut x1 = Share::zero();
-        let mut x2 = Share::zero();
-        let mut x3 = Share::zero();
-
-        match self.network.get_id() {
-            PartyID::ID0 => {
-                x1.a = a;
-                x3.b = b;
-            }
-            PartyID::ID1 => {
-                x2.a = a;
-                x1.b = b;
-            }
-            PartyID::ID2 => {
-                x3.a = a;
-                x2.b = b;
-            }
+    let shares_b = {
+        let serialized_other_share = network.receive(&prev_party, &sid).await;
+        match NetworkValue::from_network(serialized_other_share) {
+            Ok(NetworkValue::VecRing64(message)) => Ok(message),
+            _ => Err(eyre!("Error in receiving in and_many operation")),
         }
-        (x1, x2, x3)
+    }?;
+    Ok(shares_b)
+}
+
+pub(crate) async fn and_many(
+    session: &mut Session,
+    a: SliceShare<'_, u64>,
+    b: SliceShare<'_, u64>,
+) -> Result<VecShare<u64>, Error>
+where
+    Standard: Distribution<u64>,
+{
+    let shares_a = and_many_send(session, a, b).await?;
+    let shares_b = and_many_receive(session).await?;
+    let complete_shares = VecShare::from_ab(shares_a, shares_b);
+    Ok(complete_shares)
+}
+
+pub(crate) async fn transposed_pack_and(
+    session: &mut Session,
+    x1: Vec<VecShare<u64>>,
+    x2: Vec<VecShare<u64>>,
+) -> Result<Vec<VecShare<u64>>, Error>
+where
+    Standard: Distribution<u64>,
+{
+    // TODO(Dragos) this could probably be parallelized even more.
+    let mut res = Vec::with_capacity(x1.len());
+    for (x1, x2) in x1.iter().zip(x2.iter()) {
+        let shares_a = and_many_send(session, x1.as_slice(), x2.as_slice()).await?;
+        let shares_b = and_many_receive(session).await?;
+        res.push(VecShare::from_ab(shares_a, shares_b))
+    }
+    Ok(res)
+}
+
+async fn binary_add_3_get_two_carries(
+    session: &mut Session,
+    x1: Vec<VecShare<u64>>,
+    x2: Vec<VecShare<u64>>,
+    x3: Vec<VecShare<u64>>,
+    truncate_len: usize,
+) -> Result<(VecShare<Bit>, VecShare<Bit>), Error>
+where
+    Standard: Distribution<u64>,
+{
+    let len = x1.len();
+    debug_assert!(len == x2.len() && len == x3.len());
+
+    // Full adder to get 2 * c and s
+    let mut x2x3 = x2;
+    transposed_pack_xor_assign(&mut x2x3, &x3);
+    let s = transposed_pack_xor(&x1, &x2x3);
+    let mut x1x3 = x1;
+    transposed_pack_xor_assign(&mut x1x3, &x3);
+    let mut c = transposed_pack_and(session, x1x3, x2x3).await?;
+    transposed_pack_xor_assign(&mut c, &x3);
+
+    // Add 2c + s via a ripple carry adder
+    // LSB of c is 0
+    // First round: half adder can be skipped due to LSB of c being 0
+    let mut a = s;
+    let mut b = c;
+
+    // First full adder (carry is 0)
+    let mut c = and_many(session, a[1].as_slice(), b[0].as_slice()).await?;
+
+    // For last round
+    let mut b_msb = b.pop().expect("Enough elements present");
+
+    // 2 -> k
+    for (a_, b_) in a.iter_mut().skip(2).zip(b.iter_mut().skip(1)) {
+        *a_ ^= c.as_slice();
+        *b_ ^= c.as_slice();
+        let tmp_c = and_many(session, a_.as_slice(), b_.as_slice()).await?;
+        c ^= tmp_c;
     }
 
-    pub(crate) fn and_many_send<T: IntRing2k>(
-        &mut self,
-        a: SliceShare<'_, T>,
-        b: SliceShare<'_, T>,
-    ) -> Result<Vec<RingElement<T>>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        if a.len() != b.len() {
-            return Err(eyre!("InvalidSize in and_many_send"));
-        }
+    // Finally, last bit of a is 0
+    let res2 = and_many(session, b_msb.as_slice(), c.as_slice()).await?;
+    b_msb ^= c;
 
-        let mut shares_a = Vec::with_capacity(a.len());
-        for (a_, b_) in a.iter().zip(b.iter()) {
-            let rand = self.prf.gen_binary_zero_share::<T>();
-            let mut c = a_ & b_;
-            c ^= rand;
-            shares_a.push(c);
-        }
+    // Extract bits for outputs
+    let mut res1 = b_msb.convert_to_bits();
+    res1.truncate(truncate_len);
+    let mut res2 = res2.convert_to_bits();
+    res2.truncate(truncate_len);
 
-        self.network
-            .blocking_send_next_id(Utils::ring_slice_to_bytes(&shares_a))?;
-        Ok(shares_a)
-    }
+    Ok((res1, res2))
+}
 
-    pub(crate) fn and_many_receive<T: IntRing2k>(
-        &mut self,
-        shares_a: Vec<RingElement<T>>,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let len = shares_a.len();
-        let response = self.network.blocking_receive_prev_id()?;
-        let shares_b = Utils::ring_iter_from_bytes(response, len)?;
+async fn bit_inject_ot_2round_helper(
+    session: &mut Session,
+    input: VecShare<Bit>,
+) -> Result<VecShare<u16>, Error>
+where
+    Standard: Distribution<u16>,
+{
+    let len = input.len();
+    let mut wc = Vec::with_capacity(len);
+    let mut shares = VecShare::with_capacity(len);
 
-        let res = VecShare::from_avec_biter(shares_a, shares_b);
-        Ok(res)
-    }
+    for inp in input.into_iter() {
+        // new share
+        let c3 = session
+            .prf_as_mut()
+            .get_prev_prf()
+            .gen::<RingElement<u16>>();
+        shares.push(Share::new(RingElement::zero(), c3));
 
-    pub(crate) fn and_many<T: IntRing2k>(
-        &mut self,
-        a: SliceShare<'_, T>,
-        b: SliceShare<'_, T>,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let shares_a = self.and_many_send(a, b)?;
-        self.and_many_receive(shares_a)
-    }
+        // mask of the ot
+        let w0 = session
+            .prf_as_mut()
+            .get_prev_prf()
+            .gen::<RingElement<u16>>();
+        let w1 = session
+            .prf_as_mut()
+            .get_prev_prf()
+            .gen::<RingElement<u16>>();
 
-    pub(crate) fn transposed_pack_and_send<T: IntRing2k>(
-        &mut self,
-        x1: Vec<VecShare<T>>,
-        x2: Vec<VecShare<T>>,
-    ) -> Result<Vec<Vec<RingElement<T>>>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let len = x1.len();
-        debug_assert_eq!(len, x2.len());
-
-        let mut send = Vec::with_capacity(len);
-        for (x1, x2) in x1.iter().zip(x2.iter()) {
-            let send_ = self.and_many_send(x1.as_slice(), x2.as_slice())?;
-            send.push(send_);
-        }
-        Ok(send)
-    }
-
-    pub(crate) fn transposed_pack_and_receive<T: IntRing2k>(
-        &mut self,
-        shares_a: Vec<Vec<RingElement<T>>>,
-    ) -> Result<Vec<VecShare<T>>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let mut x3 = Vec::with_capacity(shares_a.len());
-
-        for shares_a_ in shares_a {
-            let res = self.and_many_receive(shares_a_)?;
-            x3.push(res);
-        }
-
-        Ok(x3)
-    }
-
-    pub(crate) fn transposed_pack_and<T: IntRing2k>(
-        &mut self,
-        x1: Vec<VecShare<T>>,
-        x2: Vec<VecShare<T>>,
-    ) -> Result<Vec<VecShare<T>>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let send = self.transposed_pack_and_send(x1, x2)?;
-        self.transposed_pack_and_receive(send)
-    }
-
-    pub(crate) fn transposed_pack_xor<T: IntRing2k>(
-        x1: &[VecShare<T>],
-        x2: &[VecShare<T>],
-    ) -> Vec<VecShare<T>> {
-        let len = x1.len();
-        debug_assert_eq!(len, x2.len());
-
-        let mut res = Vec::with_capacity(len);
-        for (x1, x2) in x1.iter().zip(x2.iter()) {
-            res.push(x1.as_slice() ^ x2.as_slice());
-        }
-        res
-    }
-
-    pub(crate) fn transposed_pack_xor_assign<T: IntRing2k>(
-        x1: &mut [VecShare<T>],
-        x2: &[VecShare<T>],
-    ) {
-        let len = x1.len();
-        debug_assert_eq!(len, x2.len());
-
-        for (x1, x2) in x1.iter_mut().zip(x2.iter()) {
-            *x1 ^= x2.as_slice();
+        let choice = inp.get_a().convert().convert();
+        if choice {
+            wc.push(w1);
+        } else {
+            wc.push(w0);
         }
     }
+    let network = session.network().clone();
+    let next_id = session.next_identity()?;
+    let sid = session.session_id();
+    network
+        .send(NetworkValue::VecRing16(wc).to_network(), &next_id, &sid)
+        .await?;
 
-    fn bit_inject_ot_2round_sender<T: IntRing2k>(
-        &mut self,
-        input: VecShare<Bit>,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let len = input.len();
-        let mut m0 = Vec::with_capacity(len);
-        let mut m1 = Vec::with_capacity(len);
-        let mut shares = VecShare::with_capacity(len);
-
-        for inp in input.into_iter() {
-            let (a, b) = inp.get_ab();
-            // new shares
-            let (c3, c2) = self.prf.gen_rands::<RingElement<T>>();
-            // mask of the ot
-            let w0 = self.prf.get_my_prf().gen::<RingElement<T>>();
-            let w1 = self.prf.get_my_prf().gen::<RingElement<T>>();
-
-            shares.push(Share::new(c3, c2));
-            let c = c3 + c2;
-            let xor = RingElement(T::from((a ^ b).convert().convert()));
-            let m0_ = xor - c;
-            let m1_ = (xor ^ RingElement::one()) - c;
-            m0.push(m0_ ^ w0);
-            m1.push(m1_ ^ w1);
+    let network = session.network().clone();
+    let next_id = session.next_identity()?;
+    let sid = session.session_id();
+    let c1 = {
+        let reply = network.receive(&next_id, &sid).await;
+        match NetworkValue::from_network(reply) {
+            Ok(NetworkValue::VecRing16(val)) => Ok(val),
+            _ => Err(eyre!("Could not deserialize properly in bit inject")),
         }
-        self.network
-            .blocking_send_prev_id(Utils::ring_slice_to_bytes(&m0))?;
-        self.network
-            .blocking_send_prev_id(Utils::ring_slice_to_bytes(&m1))?;
+    }?;
 
-        Ok(shares)
+    // Receive Reshare
+    for (s, c1) in shares.iter_mut().zip(c1) {
+        s.a = c1;
     }
+    Ok(shares)
+}
 
-    fn bit_inject_ot_2round_receiver<T: IntRing2k>(
-        &mut self,
-        input: VecShare<Bit>,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let len = input.len();
-        let m0_bytes = self.network.blocking_receive_next_id()?;
-        let m1_bytes = self.network.blocking_receive_next_id()?;
-        let wc_bytes = self.network.blocking_receive_prev_id()?;
+async fn bit_inject_ot_2round_receiver(
+    session: &mut Session,
+    input: VecShare<Bit>,
+) -> Result<VecShare<u16>, Error> {
+    let network = session.network().clone();
+    let next_id = session.next_identity()?;
+    let prev_id = session.prev_identity()?;
+    let sid = session.session_id();
 
-        let m0 = Utils::ring_iter_from_bytes(m0_bytes, len)?;
-        let m1 = Utils::ring_iter_from_bytes(m1_bytes, len)?;
-
-        let wc = Utils::ring_iter_from_bytes(wc_bytes, len)?;
-
-        let mut shares = VecShare::with_capacity(len);
-        let mut send = Vec::with_capacity(len);
-
-        for ((inp, wc), (m0, m1)) in input.into_iter().zip(wc).zip(m0.zip(m1)) {
-            // new share
-            let c2 = self.prf.get_my_prf().gen::<RingElement<T>>();
-
-            let choice = inp.get_b().convert().convert();
-            let xor = if choice { wc ^ m1 } else { wc ^ m0 };
-
-            send.push(xor);
-            shares.push(Share::new(c2, xor));
-        }
-
-        // Reshare to Helper
-        self.network
-            .blocking_send_prev_id(Utils::ring_slice_to_bytes(&send))?;
-
-        Ok(shares)
-    }
-
-    fn bit_inject_ot_2round_helper<T: IntRing2k>(
-        &mut self,
-        input: VecShare<Bit>,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let len = input.len();
-        let mut wc = Vec::with_capacity(len);
-        let mut shares = VecShare::with_capacity(len);
-
-        for inp in input.into_iter() {
-            // new share
-            let c3 = self.prf.get_prev_prf().gen::<RingElement<T>>();
-            shares.push(Share::new(RingElement::zero(), c3));
-
-            // mask of the ot
-            let w0 = self.prf.get_prev_prf().gen::<RingElement<T>>();
-            let w1 = self.prf.get_prev_prf().gen::<RingElement<T>>();
-
-            let choice = inp.get_a().convert().convert();
-            if choice {
-                wc.push(w1);
-            } else {
-                wc.push(w0);
-            }
-        }
-        self.network
-            .blocking_send_next_id(Utils::ring_slice_to_bytes(&wc))?;
-
-        // Receive Reshare
-        let c1_bytes = self.network.blocking_receive_next_id()?;
-        let c1 = Utils::ring_iter_from_bytes(c1_bytes, len)?;
-
-        for (s, c1) in shares.iter_mut().zip(c1) {
-            s.a = c1;
-        }
-        Ok(shares)
-    }
-
-    // TODO this is inbalanced, so a real implementation should actually rotate
-    // parties around
-    pub(crate) fn bit_inject_ot_2round<T: IntRing2k>(
-        &mut self,
-        input: VecShare<Bit>,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let res = match self.get_party_id() {
-            PartyID::ID0 => {
-                // OT Helper
-                self.bit_inject_ot_2round_helper(input)?
-            }
-            PartyID::ID1 => {
-                // OT Receiver
-                self.bit_inject_ot_2round_receiver(input)?
-            }
-            PartyID::ID2 => {
-                // OT Sender
-                self.bit_inject_ot_2round_sender(input)?
-            }
+    let (m0, m1, wc) = tokio::spawn(async move {
+        let reply_m0 = network.receive(&next_id, &sid).await;
+        let m0 = match NetworkValue::from_network(reply_m0) {
+            Ok(NetworkValue::VecRing16(val)) => Ok(val),
+            _ => Err(eyre!("Could not deserialize properly in bit inject")),
         };
-        Ok(res)
-    }
 
-    // TODO a dedicated bitextraction for just one element would be more
-    // efficient
-    pub fn single_extract_msb_u16<const K: usize>(
-        &mut self,
-        x: Share<u16>,
-    ) -> Result<Share<Bit>, Error> {
-        let (a, b) = self
-            .extract_msb_u16::<{ u16::BITS as usize }>(VecShare::new_vec(vec![x]))?
-            .get_at(0)
-            .get_ab();
+        let reply_m1 = network.receive(&next_id, &sid).await;
+        let m1 = match NetworkValue::from_network(reply_m1) {
+            Ok(NetworkValue::VecRing16(val)) => Ok(val),
+            _ => Err(eyre!("Could not deserialize properly in bit inject")),
+        };
 
-        Ok(Share::new(a.get_bit_as_bit(0), b.get_bit_as_bit(0)))
-    }
+        let reply_wc = network.receive(&prev_id, &sid).await;
+        let wc = match NetworkValue::from_network(reply_wc) {
+            Ok(NetworkValue::VecRing16(val)) => Ok(val),
+            _ => Err(eyre!("Could not deserialize properly in bit inject")),
+        };
+        (m0, m1, wc)
+    })
+    .await?;
 
-    // TODO a dedicated bitextraction for just one element would be more
-    // efficient
-    pub fn single_extract_msb_u32<const K: usize>(
-        &mut self,
-        x: Share<u32>,
-    ) -> Result<Share<Bit>, Error> {
-        let (a, b) = self
-            .extract_msb_u32::<{ u32::BITS as usize }>(VecShare::new_vec(vec![x]))?
-            .get_at(0)
-            .get_ab();
+    let (m0, m1, wc) = (m0?, m1?, wc?);
 
-        Ok(Share::new(a.get_bit_as_bit(0), b.get_bit_as_bit(0)))
-    }
+    let len = input.len();
+    let mut shares = VecShare::with_capacity(len);
+    let mut send = Vec::with_capacity(len);
 
-    pub fn extract_msb_u16<const K: usize>(
-        &mut self,
-        x_: VecShare<u16>,
-    ) -> Result<VecShare<u64>, Error> {
-        // let truncate_len = x_.len();
-        let x = x_.transpose_pack_u64_with_len::<K>();
-        self.extract_msb::<K>(x)
-    }
-
-    pub fn extract_msb_u32<const K: usize>(
-        &mut self,
-        x_: VecShare<u32>,
-    ) -> Result<VecShare<u64>, Error> {
-        // let truncate_len = x_.len();
-        let x = x_.transpose_pack_u64_with_len::<K>();
-        self.extract_msb::<K>(x)
-    }
-
-    // Extracts bit at position K
-    fn extract_msb<const K: usize>(
-        &mut self,
-        x: Vec<VecShare<u64>>,
-    ) -> Result<VecShare<u64>, Error> {
-        let len = x.len();
-
-        let mut x1 = Vec::with_capacity(len);
-        let mut x2 = Vec::with_capacity(len);
-        let mut x3 = Vec::with_capacity(len);
-
-        for x_ in x.into_iter() {
-            let len_ = x_.len();
-            let mut x1_ = VecShare::with_capacity(len_);
-            let mut x2_ = VecShare::with_capacity(len_);
-            let mut x3_ = VecShare::with_capacity(len_);
-            for x__ in x_.into_iter() {
-                let (x1__, x2__, x3__) = self.a2b_pre(x__);
-                x1_.push(x1__);
-                x2_.push(x2__);
-                x3_.push(x3__);
-            }
-            x1.push(x1_);
-            x2.push(x2_);
-            x3.push(x3_);
-        }
-
-        self.binary_add_3_get_msb(x1, x2, x3)
-    }
-
-    fn binary_add_3_get_two_carries<T: IntRing2k>(
-        &mut self,
-        x1: Vec<VecShare<T>>,
-        x2: Vec<VecShare<T>>,
-        x3: Vec<VecShare<T>>,
-        truncate_len: usize,
-    ) -> Result<(VecShare<Bit>, VecShare<Bit>), Error>
-    where
-        Standard: Distribution<T>,
+    for ((inp, wc), (m0, m1)) in input
+        .into_iter()
+        .zip(wc.into_iter())
+        .zip(m0.into_iter().zip(m1.into_iter()))
     {
-        let len = x1.len();
-        debug_assert!(len == x2.len() && len == x3.len());
+        // new share
+        let c2 = session.prf_as_mut().get_my_prf().gen::<RingElement<u16>>();
 
-        // Full adder to get 2 * c and s
-        let mut x2x3 = x2;
-        Self::transposed_pack_xor_assign(&mut x2x3, &x3);
-        let s = Self::transposed_pack_xor(&x1, &x2x3);
-        let mut x1x3 = x1;
-        Self::transposed_pack_xor_assign(&mut x1x3, &x3);
-        let mut c = self.transposed_pack_and(x1x3, x2x3)?;
-        Self::transposed_pack_xor_assign(&mut c, &x3);
+        let choice = inp.get_b().convert().convert();
+        let xor = if choice { wc ^ m1 } else { wc ^ m0 };
 
-        // Add 2c + s via a ripple carry adder
-        // LSB of c is 0
-        // First round: half adder can be skipped due to LSB of c being 0
-        let mut a = s;
-        let mut b = c;
+        send.push(xor);
+        shares.push(Share::new(c2, xor));
+    }
 
-        // First full adder (carry is 0)
-        let mut c = self.and_many(a[1].as_slice(), b[0].as_slice())?;
+    let network = session.network().clone();
+    let prev_id = session.prev_identity()?;
+    let sid = session.session_id();
+    // Reshare to Helper
+    network
+        .send(NetworkValue::VecRing16(send).to_network(), &prev_id, &sid)
+        .await?;
 
-        // For last round
-        let mut b_msb = b.pop().expect("Enough elements present");
+    Ok(shares)
+}
 
-        // 2 -> k
-        for (a_, b_) in a.iter_mut().skip(2).zip(b.iter_mut().skip(1)) {
-            *a_ ^= c.as_slice();
-            *b_ ^= c.as_slice();
-            let tmp_c = self.and_many(a_.as_slice(), b_.as_slice())?;
-            c ^= tmp_c;
+async fn bit_inject_ot_2round_sender(
+    session: &mut Session,
+    input: VecShare<Bit>,
+) -> Result<VecShare<u16>, Error> {
+    let len = input.len();
+    let mut m0 = Vec::with_capacity(len);
+    let mut m1 = Vec::with_capacity(len);
+    let mut shares = VecShare::with_capacity(len);
+
+    for inp in input.into_iter() {
+        let (a, b) = inp.get_ab();
+        // new shares
+        let (c3, c2) = session.prf_as_mut().gen_rands::<RingElement<u16>>();
+        // mask of the ot
+        let w0 = session.prf_as_mut().get_my_prf().gen::<RingElement<u16>>();
+        let w1 = session.prf_as_mut().get_my_prf().gen::<RingElement<u16>>();
+
+        shares.push(Share::new(c3, c2));
+        let c = c3 + c2;
+        let xor = RingElement(u16::from((a ^ b).convert().convert()));
+        let m0_ = xor - c;
+        let m1_ = (xor ^ RingElement::one()) - c;
+        m0.push(m0_ ^ w0);
+        m1.push(m1_ ^ w1);
+    }
+
+    let network = session.network().clone();
+    let prev_id = session.prev_identity()?;
+    let sid = session.session_id();
+    // TODO(Dragos) Note this can be compressed in a single round.
+    // Reshare to Helper
+    tokio::spawn(async move {
+        let _ = network
+            .send(NetworkValue::VecRing16(m0).to_network(), &prev_id, &sid)
+            .await;
+        let _ = network
+            .send(NetworkValue::VecRing16(m1).to_network(), &prev_id, &sid)
+            .await;
+    })
+    .await?;
+    Ok(shares)
+}
+
+// TODO this is inbalanced, so a real implementation should actually rotate
+// parties around
+pub(crate) async fn bit_inject_ot_2round(
+    session: &mut Session,
+    input: VecShare<Bit>,
+) -> Result<VecShare<u16>, Error> {
+    let res = match session.own_role()?.zero_based() {
+        0 => {
+            // OT Helper
+            bit_inject_ot_2round_helper(session, input).await?
         }
-
-        // Finally, last bit of a is 0
-        let res2 = self.and_many(b_msb.as_slice(), c.as_slice())?;
-        b_msb ^= c;
-
-        // Extract bits for outputs
-        let mut res1 = b_msb.convert_to_bits();
-        res1.truncate(truncate_len);
-        let mut res2 = res2.convert_to_bits();
-        res2.truncate(truncate_len);
-
-        Ok((res1, res2))
-    }
-
-    pub(crate) fn binary_add_3_get_msb<T: IntRing2k>(
-        &mut self,
-        x1: Vec<VecShare<T>>,
-        x2: Vec<VecShare<T>>,
-        mut x3: Vec<VecShare<T>>,
-        // truncate_len: usize,
-    ) -> Result<VecShare<T>, Error>
-    where
-        Standard: Distribution<T>,
-    {
-        let len = x1.len();
-        debug_assert!(len == x2.len() && len == x3.len());
-
-        // Full adder to get 2 * c and s
-        let mut x2x3 = x2;
-        Self::transposed_pack_xor_assign(&mut x2x3, &x3);
-        let s = Self::transposed_pack_xor(&x1, &x2x3);
-        let mut x1x3 = x1;
-        Self::transposed_pack_xor_assign(&mut x1x3, &x3);
-        // 2 * c
-        x1x3.pop().expect("Enough elements present");
-        x2x3.pop().expect("Enough elements present");
-        x3.pop().expect("Enough elements present");
-        let mut c = self.transposed_pack_and(x1x3, x2x3)?;
-        Self::transposed_pack_xor_assign(&mut c, &x3);
-
-        // Add 2c + s via a ripple carry adder
-        // LSB of c is 0
-        // First round: half adder can be skipped due to LSB of c being 0
-        let mut a = s;
-        let mut b = c;
-
-        // First full adder (carry is 0)
-        let mut c = self.and_many(a[1].as_slice(), b[0].as_slice())?;
-
-        // For last round
-        let mut a_msb = a.pop().expect("Enough elements present");
-        let b_msb = b.pop().expect("Enough elements present");
-
-        // 2 -> k-1
-        for (a_, b_) in a.iter_mut().skip(2).zip(b.iter_mut().skip(1)) {
-            *a_ ^= c.as_slice();
-            *b_ ^= c.as_slice();
-            let tmp_c = self.and_many(a_.as_slice(), b_.as_slice())?;
-            c ^= tmp_c;
+        1 => {
+            // OT Receiver
+            bit_inject_ot_2round_receiver(session, input).await?
         }
-
-        a_msb ^= b_msb;
-        a_msb ^= c;
-
-        // Extract bits for outputs
-        let res = a_msb;
-        // let mut res = a_msb.convert_to_bits();
-        // res.truncate(truncate_len);
-
-        Ok(res)
-    }
-
-    pub(crate) fn transposed_padded_len(len: usize) -> usize {
-        let padded_len = (len + 63) / 64;
-        padded_len * 64
-    }
-
-    pub(crate) fn lift<const K: usize>(
-        &mut self,
-        shares: VecShare<u16>,
-    ) -> Result<VecShare<u32>, Error> {
-        let len = shares.len();
-        let padded_len = Self::transposed_padded_len(len);
-
-        let mut x_a = VecShare::with_capacity(padded_len);
-        for share in shares.iter() {
-            x_a.push(Share::new(
-                RingElement(share.a.0 as u32),
-                RingElement(share.b.0 as u32),
-            ));
+        2 => {
+            // OT Sender
+            bit_inject_ot_2round_sender(session, input).await?
         }
-
-        let x = shares.transpose_pack_u64();
-
-        let len_ = x.len();
-        let mut x1 = Vec::with_capacity(len_);
-        let mut x2 = Vec::with_capacity(len_);
-        let mut x3 = Vec::with_capacity(len_);
-
-        for x_ in x.into_iter() {
-            let len__ = x_.len();
-            let mut x1_ = VecShare::with_capacity(len__);
-            let mut x2_ = VecShare::with_capacity(len__);
-            let mut x3_ = VecShare::with_capacity(len__);
-            for x__ in x_.into_iter() {
-                let (x1__, x2__, x3__) = self.a2b_pre(x__);
-                x1_.push(x1__);
-                x2_.push(x2__);
-                x3_.push(x3__);
-            }
-            x1.push(x1_);
-            x2.push(x2_);
-            x3.push(x3_);
+        _ => {
+            return Err(eyre!(
+                "Cannot deal with roles outside of the set [0, 1, 2] in bit_inject_ot"
+            ))
         }
+    };
+    Ok(res)
+}
 
-        let (mut b1, b2) = self.binary_add_3_get_two_carries(x1, x2, x3, len)?;
+pub(crate) fn mul_lift_2k<const K: u64>(val: &Share<u16>) -> Share<u32>
+where
+    u32: From<u16>,
+{
+    let a = (u32::from(val.a.0)) << K;
+    let b = (u32::from(val.b.0)) << K;
+    Share::new(RingElement(a), RingElement(b))
+}
 
-        b1.extend(b2);
+pub(crate) fn mul_lift_2k_many<const K: u64>(vals: SliceShare<u16>) -> VecShare<u32> {
+    VecShare::new_vec(vals.iter().map(mul_lift_2k::<K>).collect())
+}
 
-        // We need bit1 * 2^{ShareRing::K+1} mod 2^{ShareRing::K+K}  and bit2 *
-        // 2^{ShareRing::K+1} mod 2^{ShareRing::K+K} So we inject bit to mod
-        // 2^{K-1} and mod 2^{K-2} and use the mul_lift_2k function TODO: This
-        // one is not optimized: We send too much, since we need less than K
-        // bits
-        debug_assert!(K <= 16); // otherwise u16 does not work
-        let mut b = self.bit_inject_ot_2round::<u16>(b1)?;
-        let (b1, b2) = b.split_at_mut(len);
+pub(crate) async fn lift<const K: usize>(
+    session: &mut Session,
+    shares: VecShare<u16>,
+) -> eyre::Result<VecShare<u32>> {
+    let len = shares.len();
+    let padded_len = transposed_padded_len(len);
 
-        // Make the result mod 2^{K-1} and mod 2^{K-2} (Not required since we bitextract
-        // the correct one later) Self::share_bit_mod(&mut b1, K as u32);
-        // Self::share_bit_mod(&mut b2, K as u32 - 1);
-
-        let b1 = Self::mul_lift_2k_many::<_, { u16::K as u64 }>(b1.to_slice());
-        let b2 = Self::mul_lift_2k_many::<_, { u16::K as u64 + 1 }>(b2.to_slice());
-
-        // Finally, compute the result
-        x_a.sub_assign(b1);
-        x_a.sub_assign(b2);
-        Ok(x_a)
+    let mut x_a = VecShare::with_capacity(padded_len);
+    for share in shares.iter() {
+        x_a.push(Share::new(
+            RingElement(share.a.0 as u32),
+            RingElement(share.b.0 as u32),
+        ));
     }
 
-    // Compute code_dots > a/b * mask_dots
-    // via MSB(a * mask_dots - b * code_dots)
-    pub fn compare_threshold_masked(
-        &mut self,
-        code_dot: Share<u16>,
-        mask_dot: Share<u16>,
-    ) -> Result<Share<Bit>, Error> {
-        debug_assert!(A_BITS as u64 <= B_BITS);
+    let x = shares.transpose_pack_u64();
 
-        let y = Self::mul_lift_2k::<_, B_BITS>(&code_dot);
-        // TODO a dedicated bitextraction for just one element would be more
-        // efficient
-        let mut x = self.lift::<{ B_BITS as usize }>(VecShare::new_vec(vec![mask_dot]))?;
-        debug_assert_eq!(x.len(), 1);
-        let mut x = x.pop().expect("Enough elements present");
-        x *= A as u32;
-        x -= y;
+    let len_ = x.len();
+    let mut x1 = Vec::with_capacity(len_);
+    let mut x2 = Vec::with_capacity(len_);
+    let mut x3 = Vec::with_capacity(len_);
 
-        self.single_extract_msb_u32::<{ u16::K + B_BITS as usize }>(x)
-    }
-
-    // Compute code_dots > a/b * mask_dots
-    // via MSB(a * mask_dots - b * code_dots)
-    pub fn compare_threshold_masked_many(
-        &mut self,
-        code_dots: VecShare<u16>,
-        mask_dots: VecShare<u16>,
-    ) -> Result<VecShare<u64>, Error> {
-        debug_assert!(A_BITS as u64 <= B_BITS);
-        let len = code_dots.len();
-        assert_eq!(len, mask_dots.len());
-
-        let y = Self::mul_lift_2k_many::<_, B_BITS>(code_dots.as_slice());
-        let mut x = self.lift::<{ B_BITS as usize }>(mask_dots)?;
-        for x_ in x.iter_mut() {
-            *x_ *= A as u32;
+    for x_ in x.into_iter() {
+        let len__ = x_.len();
+        let mut x1_ = VecShare::with_capacity(len__);
+        let mut x2_ = VecShare::with_capacity(len__);
+        let mut x3_ = VecShare::with_capacity(len__);
+        for x__ in x_.into_iter() {
+            let (x1__, x2__, x3__) = a2b_pre(session, x__)?;
+            x1_.push(x1__);
+            x2_.push(x2__);
+            x3_.push(x3__);
         }
-
-        x.sub_assign(y);
-        self.extract_msb_u32::<{ u16::K + B_BITS as usize }>(x)
+        x1.push(x1_);
+        x2.push(x2_);
+        x3.push(x3_);
     }
 
-    pub fn open_bit(&mut self, share: Share<Bit>) -> Result<bool, Error> {
-        let c = Utils::blocking_send_and_receive_value(&mut self.network, &share.b)?;
-        Ok((share.a ^ share.b ^ c).convert().convert())
+    let (mut b1, b2) = binary_add_3_get_two_carries(session, x1, x2, x3, len).await?;
+    b1.extend(b2);
+
+    // We need bit1 * 2^{ShareRing::K+1} mod 2^{ShareRing::K+K}  and bit2 *
+    // 2^{ShareRing::K+1} mod 2^{ShareRing::K+K} So we inject bit to mod
+    // 2^{K-1} and mod 2^{K-2} and use the mul_lift_2k function TODO: This
+    // one is not optimized: We send too much, since we need less than K
+    // bits
+    debug_assert!(K <= 16); // otherwise u16 does not work
+    let mut b = bit_inject_ot_2round(session, b1).await?;
+    let (b1, b2) = b.split_at_mut(len);
+
+    // Make the result mod 2^{K-1} and mod 2^{K-2} (Not required since we bitextract
+    // the correct one later) Self::share_bit_mod(&mut b1, K as u32);
+    // Self::share_bit_mod(&mut b2, K as u32 - 1);
+
+    let b1 = mul_lift_2k_many::<{ u16::K as u64 }>(b1.to_slice());
+    let b2 = mul_lift_2k_many::<{ u16::K as u64 + 1 }>(b2.to_slice());
+
+    // Finally, compute the result
+    x_a.sub_assign(b1);
+    x_a.sub_assign(b2);
+    Ok(x_a)
+}
+
+// MSB related code
+pub(crate) async fn binary_add_3_get_msb(
+    session: &mut Session,
+    x1: Vec<VecShare<u64>>,
+    x2: Vec<VecShare<u64>>,
+    mut x3: Vec<VecShare<u64>>,
+    // truncate_len: usize,
+) -> Result<VecShare<u64>, Error> {
+    let len = x1.len();
+    debug_assert!(len == x2.len() && len == x3.len());
+
+    // Full adder to get 2 * c and s
+    let mut x2x3 = x2;
+    transposed_pack_xor_assign(&mut x2x3, &x3);
+    let s = transposed_pack_xor(&x1, &x2x3);
+    let mut x1x3 = x1;
+    transposed_pack_xor_assign(&mut x1x3, &x3);
+    // 2 * c
+    x1x3.pop().expect("Enough elements present");
+    x2x3.pop().expect("Enough elements present");
+    x3.pop().expect("Enough elements present");
+    let mut c = transposed_pack_and(session, x1x3, x2x3).await?;
+    transposed_pack_xor_assign(&mut c, &x3);
+
+    // Add 2c + s via a ripple carry adder
+    // LSB of c is 0
+    // First round: half adder can be skipped due to LSB of c being 0
+    let mut a = s;
+    let mut b = c;
+
+    // First full adder (carry is 0)
+    let mut c = and_many(session, a[1].as_slice(), b[0].as_slice()).await?;
+
+    // For last round
+    let mut a_msb = a.pop().expect("Enough elements present");
+    let b_msb = b.pop().expect("Enough elements present");
+
+    // 2 -> k-1
+    for (a_, b_) in a.iter_mut().skip(2).zip(b.iter_mut().skip(1)) {
+        *a_ ^= c.as_slice();
+        *b_ ^= c.as_slice();
+        let tmp_c = and_many(session, a_.as_slice(), b_.as_slice()).await?;
+        c ^= tmp_c;
     }
 
-    pub fn open_bit_many(&mut self, shares: VecShare<Bit>) -> Result<Vec<bool>, Error> {
-        let shares_b = shares.iter().map(|s| &s.b);
-        let bytes = Utils::blocking_send_iter_and_receive(&mut self.network, shares_b)?;
-        let shares_c = Utils::ring_iter_from_bytes(bytes, shares.len())?;
-        let res = shares
-            .into_iter()
-            .zip(shares_c)
-            .map(|(s, c)| {
-                let (a, b) = s.get_ab();
-                (a ^ b ^ c).convert().convert()
-            })
-            .collect();
-        Ok(res)
+    a_msb ^= b_msb;
+    a_msb ^= c;
+
+    // Extract bits for outputs
+    let res = a_msb;
+    // let mut res = a_msb.convert_to_bits();
+    // res.truncate(truncate_len);
+
+    Ok(res)
+}
+
+// Extracts bit at position K
+async fn extract_msb<const K: usize>(
+    session: &mut Session,
+    x: Vec<VecShare<u64>>,
+) -> Result<VecShare<u64>, Error> {
+    let len = x.len();
+
+    let mut x1 = Vec::with_capacity(len);
+    let mut x2 = Vec::with_capacity(len);
+    let mut x3 = Vec::with_capacity(len);
+
+    for x_ in x.into_iter() {
+        let len_ = x_.len();
+        let mut x1_ = VecShare::with_capacity(len_);
+        let mut x2_ = VecShare::with_capacity(len_);
+        let mut x3_ = VecShare::with_capacity(len_);
+        for x__ in x_.into_iter() {
+            let (x1__, x2__, x3__) = a2b_pre(session, x__)?;
+            x1_.push(x1__);
+            x2_.push(x2__);
+            x3_.push(x3__);
+        }
+        x1.push(x1_);
+        x2.push(x2_);
+        x3.push(x3_);
     }
 
-    pub fn open_bin<T: IntRing2k>(&mut self, share: Share<T>) -> Result<T, Error> {
-        let c = Utils::blocking_send_and_receive_value(&mut self.network, &share.b)?;
-        Ok((share.a ^ share.b ^ c).convert())
-    }
+    binary_add_3_get_msb(session, x1, x2, x3).await
+}
 
-    pub fn open_bin_many<T: IntRing2k>(&mut self, shares: VecShare<T>) -> Result<Vec<T>, Error> {
-        let shares_b = shares.iter().map(|s| &s.b);
-        let bytes: bytes::BytesMut =
-            Utils::blocking_send_iter_and_receive(&mut self.network, shares_b)?;
-        let shares_c = Utils::ring_iter_from_bytes(bytes, shares.len())?;
-        let res = shares
-            .into_iter()
-            .zip(shares_c)
-            .map(|(s, c)| {
-                let (a, b) = s.get_ab();
-                (a ^ b ^ c).convert()
-            })
-            .collect();
-        Ok(res)
-    }
+pub async fn extract_msb_u32<const K: usize>(
+    session: &mut Session,
+    x_: VecShare<u32>,
+) -> Result<VecShare<u64>, Error> {
+    // let truncate_len = x_.len();
+    let x = x_.transpose_pack_u64_with_len::<K>();
+    extract_msb::<K>(session, x).await
+}
+
+// TODO a dedicated bitextraction for just one element would be more
+// efficient
+pub async fn single_extract_msb_u32<const K: usize>(
+    session: &mut Session,
+    x: Share<u32>,
+) -> Result<Share<Bit>, Error> {
+    let (a, b) = extract_msb_u32::<{ u32::BITS as usize }>(session, VecShare::new_vec(vec![x]))
+        .await?
+        .get_at(0)
+        .get_ab();
+
+    Ok(Share::new(a.get_bit_as_bit(0), b.get_bit_as_bit(0)))
+}
+
+pub async fn open_bin(session: &mut Session, share: Share<Bit>) -> Result<Bit, Error> {
+    // send to next_party
+    let next_party = session.next_identity()?;
+    let network = session.network().clone();
+    let sid = session.session_id();
+    let message = share.b;
+    network
+        .send(
+            NetworkValue::RingElementBit(message).to_network(),
+            &next_party,
+            &sid,
+        )
+        .await?;
+
+    // receiving from previous party
+    let network = session.network().clone();
+    let sid = session.session_id();
+    let prev_party = session.prev_identity()?;
+    let c = {
+        let serialized_other_share = network.receive(&prev_party, &sid).await;
+        match NetworkValue::from_network(serialized_other_share) {
+            Ok(NetworkValue::RingElementBit(message)) => Ok(message),
+            _ => Err(eyre!("Error in receiving in open_bin operation")),
+        }
+    }?;
+
+    // xor shares with the received share
+    Ok((share.a ^ share.b ^ c).convert())
 }
