+++ iris-mpc-cpu/src/protocol/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod binary;
-pub(crate) mod dots;
-pub(crate) mod iris_worker;
+pub mod ng_worker;
 pub(crate) mod prf;
